reorganise event defs, make instance cutscene playback generic
